+++ src/settings/index.js
@@ -1,3 +1,3 @@
 export default Object.freeze({
-  VERSION: '0.0.5'
+  VERSION: '0.0.6'
 })